NextAuth route file
