+++ app/api/auth/[...nextauth]/route.ts
@@ -1,83 +1,5 @@
 import NextAuth from "next-auth";
-import { AuthOptions } from "next-auth";
-import { compare } from "bcrypt";
-import CredentialsProvider from "next-auth/providers/credentials";
-import { prisma } from "../../../../lib/prisma";
-
-if (!process.env.NEXTAUTH_SECRET) {
-  throw new Error('Please provide process.env.NEXTAUTH_SECRET');
-}
-
-const DEMO_USER_ID = '65f3f3f3f3f3f3f3f3f3f3f3';
-
-export const authOptions: AuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" }
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          throw new Error('Missing credentials');
-        }
-
-        if (credentials.email === 'demo@example.com' && credentials.password === 'demo123') {
-          return {
-            id: DEMO_USER_ID,
-            email: 'demo@example.com',
-            name: 'Demo User',
-          };
-        }
-
-        const user = await prisma.user.findUnique({
-          where: { email: credentials.email }
-        });
-
-        if (!user) {
-          throw new Error('Invalid credentials');
-        }
-
-        const isPasswordValid = await compare(credentials.password, user.password);
-
-        if (!isPasswordValid) {
-          throw new Error('Invalid credentials');
-        }
-
-        return {
-          id: user.id,
-          email: user.email,
-        };
-      }
-    }),
-  ],
-  secret: process.env.NEXTAUTH_SECRET,
-  pages: {
-    signIn: '/auth/signin',
-  },
-  session: {
-    strategy: "jwt",
-    maxAge: 30 * 24 * 60 * 60, 
-  },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-        token.email = user.email;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.id;
-        session.user.email = token.email as string;
-      }
-      return session;
-    },
-  },
-  debug: process.env.NODE_ENV === 'development',
-};
+import { authOptions } from "../options";
 
 const handler = NextAuth(authOptions);
