+++ app/api/projects/route.ts
@@ -1,7 +1,7 @@
 import { prisma } from '../../../lib/prisma';
 import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../auth/[...nextauth]/route';
+import { authOptions } from '../auth/options';
 
 export async function GET() {
   try {
